Tentativa de configurar elastic search
