+++ Usuario/API/Middleware/ConfigureElasticMiddleware.cs
@@ -1,0 +1,49 @@
+﻿using Serilog.Events;
+using Serilog.Sinks.Elasticsearch;
+using Serilog;
+using System.Reflection;
+
+namespace Usuario.API.Middleware
+{
+    public class ConfigureElasticMiddleware
+    {
+        private readonly RequestDelegate _next;
+        public ConfigureElasticMiddleware(RequestDelegate next)
+        {
+            _next = next;
+            ConfigureLogging();
+        }
+
+        public async Task Invoke(HttpContext context)
+        {
+            await _next(context);
+        }
+
+        private static void ConfigureLogging()
+        {
+            var environment = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT");
+            var configuration = new ConfigurationBuilder()
+                .SetBasePath(Directory.GetCurrentDirectory())
+                .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
+                .Build();
+
+            string url = configuration.GetSection("ElasticConfiguration:Uri").Value!;
+
+            Log.Logger = new LoggerConfiguration()
+                .MinimumLevel.Override("Microsoft", LogEventLevel.Information)
+                .Enrich.FromLogContext()
+                .Enrich.WithEnvironmentName()
+                .Enrich.WithMachineName()
+                .WriteTo.Console()
+                .WriteTo.Debug()
+                .WriteTo.Elasticsearch(new ElasticsearchSinkOptions(new Uri(url))
+                {
+                    AutoRegisterTemplate = true,
+                    AutoRegisterTemplateVersion = AutoRegisterTemplateVersion.ESv6,
+                    IndexFormat = $"{Assembly.GetExecutingAssembly().GetName().Name!.ToLower().Replace(".", "-")}-{environment?.ToLower().Replace(".", "-")}-{DateTime.UtcNow:yyyy-MM}"
+                })
+                .ReadFrom.Configuration(configuration)
+                .CreateLogger();
+        }
+    }
+}
