+++ Usuario/Infra/Repositories/UserRepository.cs
@@ -17,29 +17,29 @@ namespace Usuario.Infra.Repositories
 
         public User Login(Login login)
         {
-            var connection = _connection.getConnection();
-            connection.Open();
+            //var connection = _connection.getConnection();
+            //connection.Open();
 
-            try
-            {
-                var sql = "Select id, login, password from login where login = @login and password = @password";
+            //try
+            //{
+            //    var sql = "Select id, login, password from login where login = @login and password = @password";
 
-                var parameters = new DynamicParameters();
-                parameters.Add("@login", login.login);
-                parameters.Add("@password", login.password);
+            //    var parameters = new DynamicParameters();
+            //    parameters.Add("@login", login.login);
+            //    parameters.Add("@password", login.password);
 
-                var user = connection.QueryFirst<User>(sql, parameters);
+            //    var user = connection.QueryFirst<User>(sql, parameters);
 
-                return user;
-            }
-            catch (Exception)
-            {
-                throw;
-            }
-            finally
-            {
-                connection.Close();
-            }
+                return new User();
+            //}
+            //catch (Exception)
+            //{
+            //    throw;
+            //}
+            //finally
+            //{
+            //    connection.Close();
+            //}
         }
     }
 }
