+++ Usuario/Program.cs
@@ -1,3 +1,7 @@
+using Serilog;
+using Serilog.Events;
+using Serilog.Sinks.Elasticsearch;
+using System.Reflection;
 using Usuario.API.Middleware;
 using Usuario.API.Routes;
 using Usuario.Application.IoC;
@@ -14,6 +18,33 @@ builder.Services.AddInfraDependencyResolver();
 
 var app = builder.Build();
 
+var environment = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT");
+var configuration = new ConfigurationBuilder()
+    .SetBasePath(Directory.GetCurrentDirectory())
+    .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
+    .Build();
+
+string url = configuration.GetSection("ElasticConfiguration:Uri").Value!;
+
+Log.Logger = new LoggerConfiguration()
+    .MinimumLevel.Override("Microsoft", LogEventLevel.Information)
+    .Enrich.FromLogContext()
+    .Enrich.WithEnvironmentName()
+    .Enrich.WithMachineName()
+    .WriteTo.Console()
+    .WriteTo.Debug()
+    .WriteTo.Elasticsearch(new ElasticsearchSinkOptions(new Uri(url))
+    {
+        AutoRegisterTemplate = true,
+        AutoRegisterTemplateVersion = AutoRegisterTemplateVersion.ESv6,
+        IndexFormat = $"{Assembly.GetExecutingAssembly().GetName().Name!.ToLower().Replace(".", "-")}-{environment?.ToLower().Replace(".", "-")}-{DateTime.UtcNow:yyyy-MM}"
+    })
+    .ReadFrom.Configuration(configuration)
+    .CreateLogger();
+
+Log.Information("teste");
+Log.CloseAndFlush();
+
 // Configure the HTTP request pipeline.
 if (app.Environment.IsDevelopment())
 {
@@ -24,7 +55,9 @@ if (app.Environment.IsDevelopment())
 app.UseHttpsRedirection();
 app.AuthenticationEndpoints();
 
+//app.UseMiddleware<ConfigureElasticMiddleware>();
 app.UseMiddleware<LogMiddleware>();
 app.UseMiddleware<ErrorMiddleware>();
 
 app.Run();
+
